More Goodies
Created mel and mfcc analyzer functions that can be run from main on the entire audio directory.

both analyzers create and save charts as png's. MFCC also creates and saves the table in the new tables directory. MFCC analyzer is a riff on Brandons riff on Valerio Velardo's original notebook

Added a new MFCC notebook that is more interactive with a few extra data goodies including a heatmap.

Minor updates to the Mel notebook